requirements.txt
# Web Framework
fastapi>=0.104.0
uvicorn[standard]>=0.24.0

# AI/ML
openai>=1.12.0
langchain>=0.1.0
transformers>=4.36.0
torch>=2.1.0

# Data Processing
pandas>=2.1.0
numpy>=1.24.0
pyarrow>=14.0.0
polars>=0.20.0
requests>=2.31.0
httpx>=0.25.0

# Database
sqlalchemy>=2.0.0
alembic>=1.12.0

# Character Intelligence
pydantic>=2.5.0
pydantic-settings>=2.1.0
python-multipart>=0.0.6

# Social Media APIs
tweepy>=4.14.0
instagrapi>=2.0.0

# Testing & Quality
pytest>=7.4.0
black>=23.0.0
flake8>=6.0.0
mypy>=1.6.0
typer>=0.12.0

# Metrics & Stats
scipy>=1.11.0
statsmodels>=0.14.0
scikit-learn>=1.3.0
matplotlib>=3.8.0

# Config & Crypto
PyYAML>=6.0.1
cryptography>=41.0.0

# Environment
python-dotenv>=1.0.0
beautifulsoup4==4.12.3